feature: metrics container
